+++ examples/04_fetch/App.tsx
@@ -1,15 +1,17 @@
 import * as React from 'react';
 
+import { GlobalStateProvider } from './state';
+
 import ErrorMessage from './ErrorMessage';
 import PageInfo from './PageInfo';
 import RandomButton from './RandomButton';
 
 const App = () => (
-  <div>
+  <GlobalStateProvider>
     <PageInfo />
     <RandomButton />
     <ErrorMessage />
-  </div>
+  </GlobalStateProvider>
 );
 
 export default App;