+++ examples/06_reducer/App.tsx
@@ -1,17 +1,19 @@
 import * as React from 'react';
 
+import { GlobalStateProvider } from './state';
+
 import Counter from './Counter';
 import Person from './Person';
 
 const App = () => (
-  <div>
+  <GlobalStateProvider>
     <h1>Counter</h1>
     <Counter />
     <Counter />
     <h1>Person</h1>
     <Person />
     <Person />
-  </div>
+  </GlobalStateProvider>
 );
 
 export default App;