+++ examples/09_comparison/App.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 
 import Counter from './Counter';
 import Person from './Person';
+import { GlobalStateProvider } from './state';
 
 import Counter2 from './Counter2';
 import Person2 from './Person2';
@@ -22,12 +23,14 @@ const App = () => (
     </div>
     <div>
       <h1>react-hooks-global-state</h1>
-      <h2>Counter</h2>
-      <Counter />
-      <Counter />
-      <h2>Person</h2>
-      <Person />
-      <Person />
+      <GlobalStateProvider>
+        <h2>Counter</h2>
+        <Counter />
+        <Counter />
+        <h2>Person</h2>
+        <Person />
+        <Person />
+      </GlobalStateProvider>
     </div>
   </div>
 );