+++ examples/09_comparison/state.ts
@@ -2,4 +2,4 @@ import { createStore } from '../../src/index';
 
 import { initialState, reducer } from './common';
 
-export const { dispatch, useGlobalState } = createStore(reducer, initialState);
+export const { GlobalStateProvider, dispatch, useGlobalState } = createStore(reducer, initialState);